Move TracksController into the admin namespace
